Move unit tests for is_valid_cidr, is_valid_ipv4, and get_ioc_type to… (#1380)
* Move unit tests for is_valid_cidr, is_valid_ipv4, and get_ioc_type to test_utils.py

* Move unit tests for is_valid_cidr, is_valid_ipv4, and get_ioc_type to test_utils.py
