+++ tests/test_utils.py
@@ -5,7 +5,286 @@ from ipaddress import ip_address
 
 from django.test import SimpleTestCase
 
-from greedybear.utils import is_ip_address, is_non_global_ip, is_sha256hash, is_valid_domain
+from greedybear.consts import DOMAIN, IP
+from greedybear.utils import (
+    get_ioc_type,
+    is_ip_address,
+    is_non_global_ip,
+    is_sha256hash,
+    is_valid_cidr,
+    is_valid_domain,
+    is_valid_ipv4,
+)
+
+from . import CustomTestCase
+
+
+class TestIsValidCIDR(CustomTestCase):
+    def test_valid_cidr_returns_true_and_cleaned_cidr(self):
+        is_valid, cidr = is_valid_cidr("192.168.1.0/24")
+        self.assertTrue(is_valid)
+        self.assertEqual(cidr, "192.168.1.0/24")
+
+    def test_valid_cidr_edge_cases(self):
+        is_valid, cidr = is_valid_cidr("0.0.0.0/0")
+        self.assertTrue(is_valid)
+        self.assertEqual(cidr, "0.0.0.0/0")
+
+        is_valid, cidr = is_valid_cidr("255.255.255.255/32")
+        self.assertTrue(is_valid)
+        self.assertEqual(cidr, "255.255.255.255/32")
+
+        is_valid, cidr = is_valid_cidr("10.0.0.0/8")
+        self.assertTrue(is_valid)
+        self.assertEqual(cidr, "10.0.0.0/8")
+
+    def test_cidr_with_whitespace_strips_and_validates(self):
+        is_valid, cidr = is_valid_cidr("  192.168.1.0/24")
+        self.assertTrue(is_valid)
+        self.assertEqual(cidr, "192.168.1.0/24")
+
+        is_valid, cidr = is_valid_cidr("192.168.1.0/24  ")
+        self.assertTrue(is_valid)
+        self.assertEqual(cidr, "192.168.1.0/24")
+
+        is_valid, cidr = is_valid_cidr("  192.168.1.0/24  ")
+        self.assertTrue(is_valid)
+        self.assertEqual(cidr, "192.168.1.0/24")
+
+    def test_invalid_cidr_out_of_range_octets(self):
+        invalid = [
+            "256.1.1.0/24",
+            "1.256.1.0/24",
+            "1.1.256.0/24",
+            "999.999.999.999/24",
+        ]
+
+        for value in invalid:
+            is_valid, cidr = is_valid_cidr(value)
+            self.assertFalse(is_valid)
+            self.assertIsNone(cidr)
+
+    def test_invalid_cidr_incomplete_format(self):
+        invalid = [
+            "192.168.1/24",
+            "192.168/24",
+            "192/24",
+            "/24",
+        ]
+
+        for value in invalid:
+            is_valid, cidr = is_valid_cidr(value)
+            self.assertFalse(is_valid)
+            self.assertIsNone(cidr)
+
+    def test_invalid_cidr_too_many_octets(self):
+        is_valid, cidr = is_valid_cidr("1.2.3.4.5/24")
+        self.assertFalse(is_valid)
+        self.assertIsNone(cidr)
+
+    def test_invalid_cidr_domains(self):
+        is_valid, cidr = is_valid_cidr("example.com/24")
+        self.assertFalse(is_valid)
+        self.assertIsNone(cidr)
+
+        is_valid, cidr = is_valid_cidr("sub.example.com/16")
+        self.assertFalse(is_valid)
+        self.assertIsNone(cidr)
+
+    def test_invalid_cidr_ipv6_addresses(self):
+        is_valid, cidr = is_valid_cidr("2001:db8::/32")
+        self.assertFalse(is_valid)
+        self.assertIsNone(cidr)
+
+        is_valid, cidr = is_valid_cidr("::1/128")
+        self.assertFalse(is_valid)
+        self.assertIsNone(cidr)
+
+    def test_invalid_cidr_random_strings(self):
+        is_valid, cidr = is_valid_cidr("/w00tw00t.at.ISC.SANS.DFind:)")
+        self.assertFalse(is_valid)
+        self.assertIsNone(cidr)
+
+        is_valid, cidr = is_valid_cidr("not a cidr")
+        self.assertFalse(is_valid)
+        self.assertIsNone(cidr)
+
+        is_valid, cidr = is_valid_cidr("")
+        self.assertFalse(is_valid)
+        self.assertIsNone(cidr)
+
+    def test_invalid_cidr_special_characters(self):
+        is_valid, cidr = is_valid_cidr("192.168.1.0/24#comment")
+        self.assertFalse(is_valid)
+        self.assertIsNone(cidr)
+
+        is_valid, cidr = is_valid_cidr("192.168.1.0/24 # comment")
+        self.assertFalse(is_valid)
+        self.assertIsNone(cidr)
+
+        is_valid, cidr = is_valid_cidr("10.0.0.0/8 some text")
+        self.assertFalse(is_valid)
+        self.assertIsNone(cidr)
+
+    def test_invalid_cidr_negative_numbers(self):
+        invalid = [
+            "-1.1.1.1/24",
+            "192.168.1.0/-1",
+            "192.168.1.0/33",
+        ]
+
+        for value in invalid:
+            is_valid, cidr = is_valid_cidr(value)
+            self.assertFalse(is_valid)
+            self.assertIsNone(cidr)
+
+
+class TestIsValidIpv4(CustomTestCase):
+    def test_valid_ipv4_returns_true_and_cleaned_ip(self):
+        is_valid, ip = is_valid_ipv4("1.2.3.4")
+        self.assertTrue(is_valid)
+        self.assertEqual(ip, "1.2.3.4")
+
+    def test_valid_ipv4_edge_cases(self):
+        # Test boundary values
+        is_valid, ip = is_valid_ipv4("0.0.0.0")
+        self.assertTrue(is_valid)
+        self.assertEqual(ip, "0.0.0.0")
+
+        is_valid, ip = is_valid_ipv4("255.255.255.255")
+        self.assertTrue(is_valid)
+        self.assertEqual(ip, "255.255.255.255")
+
+        is_valid, ip = is_valid_ipv4("192.168.1.1")
+        self.assertTrue(is_valid)
+        self.assertEqual(ip, "192.168.1.1")
+
+    def test_ipv4_with_whitespace_strips_and_validates(self):
+        # Test leading whitespace
+        is_valid, ip = is_valid_ipv4("  1.2.3.4")
+        self.assertTrue(is_valid)
+        self.assertEqual(ip, "1.2.3.4")
+
+        # Test trailing whitespace
+        is_valid, ip = is_valid_ipv4("1.2.3.4  ")
+        self.assertTrue(is_valid)
+        self.assertEqual(ip, "1.2.3.4")
+
+        # Test both
+        is_valid, ip = is_valid_ipv4("  1.2.3.4  ")
+        self.assertTrue(is_valid)
+        self.assertEqual(ip, "1.2.3.4")
+
+    def test_invalid_ipv4_out_of_range_octets(self):
+        # Test octets > 255
+        is_valid, ip = is_valid_ipv4("256.1.1.1")
+        self.assertFalse(is_valid)
+        self.assertIsNone(ip)
+
+        is_valid, ip = is_valid_ipv4("1.256.1.1")
+        self.assertFalse(is_valid)
+        self.assertIsNone(ip)
+
+        is_valid, ip = is_valid_ipv4("1.1.256.1")
+        self.assertFalse(is_valid)
+        self.assertIsNone(ip)
+
+        is_valid, ip = is_valid_ipv4("1.1.1.256")
+        self.assertFalse(is_valid)
+        self.assertIsNone(ip)
+
+        is_valid, ip = is_valid_ipv4("999.999.999.999")
+        self.assertFalse(is_valid)
+        self.assertIsNone(ip)
+
+    def test_invalid_ipv4_incomplete_format(self):
+        # Too few octets
+        is_valid, ip = is_valid_ipv4("1.2.3")
+        self.assertFalse(is_valid)
+        self.assertIsNone(ip)
+
+        is_valid, ip = is_valid_ipv4("1.2")
+        self.assertFalse(is_valid)
+        self.assertIsNone(ip)
+
+        is_valid, ip = is_valid_ipv4("1")
+        self.assertFalse(is_valid)
+        self.assertIsNone(ip)
+
+    def test_invalid_ipv4_too_many_octets(self):
+        is_valid, ip = is_valid_ipv4("1.2.3.4.5")
+        self.assertFalse(is_valid)
+        self.assertIsNone(ip)
+
+    def test_invalid_ipv4_domains(self):
+        is_valid, ip = is_valid_ipv4("example.com")
+        self.assertFalse(is_valid)
+        self.assertIsNone(ip)
+
+        is_valid, ip = is_valid_ipv4("sub.example.com")
+        self.assertFalse(is_valid)
+        self.assertIsNone(ip)
+
+    def test_invalid_ipv4_ipv6_addresses(self):
+        # IPv6 should not be valid for IPv4 validation
+        is_valid, ip = is_valid_ipv4("2001:0db8:85a3::8a2e:0370:7334")
+        self.assertFalse(is_valid)
+        self.assertIsNone(ip)
+
+        is_valid, ip = is_valid_ipv4("::1")
+        self.assertFalse(is_valid)
+        self.assertIsNone(ip)
+
+    def test_invalid_ipv4_random_strings(self):
+        is_valid, ip = is_valid_ipv4("/w00tw00t.at.ISC.SANS.DFind:)")
+        self.assertFalse(is_valid)
+        self.assertIsNone(ip)
+
+        is_valid, ip = is_valid_ipv4("not an ip")
+        self.assertFalse(is_valid)
+        self.assertIsNone(ip)
+
+        is_valid, ip = is_valid_ipv4("")
+        self.assertFalse(is_valid)
+        self.assertIsNone(ip)
+
+    def test_invalid_ipv4_special_characters(self):
+        is_valid, ip = is_valid_ipv4("1.2.3.4#comment")
+        self.assertFalse(is_valid)
+        self.assertIsNone(ip)
+
+        is_valid, ip = is_valid_ipv4("1.2.3.4 # comment")
+        self.assertFalse(is_valid)
+        self.assertIsNone(ip)
+
+    def test_invalid_ipv4_negative_numbers(self):
+        is_valid, ip = is_valid_ipv4("-1.2.3.4")
+        self.assertFalse(is_valid)
+        self.assertIsNone(ip)
+
+        is_valid, ip = is_valid_ipv4("1.-2.3.4")
+        self.assertFalse(is_valid)
+        self.assertIsNone(ip)
+
+
+class TestGetIocType(CustomTestCase):
+    def test_ipv4_returns_ip(self):
+        self.assertEqual(get_ioc_type("1.2.3.4"), IP)
+
+    def test_ipv4_edge_cases(self):
+        self.assertEqual(get_ioc_type("0.0.0.0"), IP)
+        self.assertEqual(get_ioc_type("255.255.255.255"), IP)
+        self.assertEqual(get_ioc_type("192.168.1.1"), IP)
+
+    def test_domain_returns_domain(self):
+        self.assertEqual(get_ioc_type("example.com"), DOMAIN)
+
+    def test_subdomain_returns_domain(self):
+        self.assertEqual(get_ioc_type("sub.example.com"), DOMAIN)
+
+    def test_invalid_ip_returns_domain(self):
+        self.assertEqual(get_ioc_type("256.1.1.1"), DOMAIN)
+        self.assertEqual(get_ioc_type("1.2.3"), DOMAIN)
 
 
 class UtilsTestCase(SimpleTestCase):
